added loss calculation #23, adjusted setup.py for additional package, moved kmeans to networkclustering.py
